Add files via upload
This project is just for learning and a hobby, but I really enjoy making stuff like this. I know the code is pretty bad, there are lots of glitches and things don’t always work right, but I’m not gonna give up on it. I’ll keep learning and slowly improve the code and how it runs.